refactor: created FAQItem
