+++ src/app/[locale]/page.jsx
@@ -1,6 +1,6 @@
 import CarouselTestSection from "@/src/components/CarouselTestSection/CarouselTestSection";
 import Counter from "@/src/components/Counter/Counter";
-
+import FAQ from "@/src/components/FAQ/FAQ";
 import RoleCardTestSection from "@/src/components/RoleCardTestSection/RoleCardTestSection";
 import Timeline from "@/src/components/Timeline/Timeline";
 
@@ -12,6 +12,7 @@ export default function Home() {
       <RoleCardTestSection />
       <Counter />
       <Timeline />
+      <FAQ />
     </main>
   );
 }
