+++ src/components/FAQ/FAQItem/FAQItem.module.scss
@@ -1,0 +1,79 @@
+@use "sass:math";
+@use "../../../styles/mixins" as mixins;
+@use "../../../styles/variables" as var;
+
+.faqItem {
+  position: relative;
+  cursor: pointer;
+  overflow: hidden;
+  @media (any-hover: none) {
+    cursor: default;
+  }
+  &::before {
+    content: "";
+    position: absolute;
+    width: 100%;
+    height: 2px;
+    bottom: 0;
+    left: 0;
+    background: linear-gradient(
+      139deg,
+      #db97c5 0%,
+      #cc4fa3 16.91%,
+      #a03a7e 39.6%,
+      #6d2656 78.5%,
+      #490735 100%
+    );
+  }
+}
+.titleRow {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 16px 0 16px 10px;
+  gap: 20px;
+}
+.title {
+  @include mixins.apply-font-styles(var.$h3-font);
+  font-size: 22px;
+  line-height: 140%;
+  font-weight: 700;
+}
+.plus {
+  width: 48px;
+  height: 48px;
+  padding: 10px;
+  & > span {
+    display: block;
+    width: 100%;
+    height: 3px;
+    border-radius: 5px;
+    background-color: #fff;
+    transition: transform var.$transition-duration-base
+      var.$transition-timing-base 0s;
+    &:first-child {
+      transform: rotate(-90deg) translateX(-3px);
+    }
+  }
+}
+.plusActive {
+  & > span {
+    &:first-child {
+      transform: rotate(0) translate(0, 3px);
+    }
+  }
+}
+
+.desc {
+  @include mixins.apply-font-styles(var.$text-16-font);
+  padding: 0 10px;
+  font-size: 18px;
+  height: 0;
+  width: calc(100% - 48px);
+  transition: height var.$transition-duration-base var.$transition-timing-base
+    0s;
+}
+
+.descActive {
+  height: 80px;
+}
